+++ CHANGELOG.md
@@ -11,6 +11,7 @@ New features:
   `MOVE` opcodes to be eliminated.
 - Improved closure creation behaviour to allow a greater set of registers and remove more unnecessary `MOVE` opcodes.
 - Added fast-paths for many prelude operations where there are only 2 operands.
+- Added some syntax detection rules into `menai-check` to improve pinpointing mismatched paren issues.
 
 Bug fixes: